split affected runtime owner
